Created class Game and player, used the class Game to store params.
-Created Game class
-Created Player class

Now the form in the index page is working. It submit the value to
'/name'. The param 'player' is then saved inside the class Game with
the help of a class method.

Added a sinatra filter 'before ' which give to each route access
to the Game class and its information.

Thanks to this technique we showed the player's name in the play view file.

Added feature test to check if the input form in the index page works.

-All tests green
-Rubocop Happy.

More detailed explanatory text, if necessary. Wrap it to about 72
characters or so. In some contexts, the first line is treated as the
subject of the commit and the rest of the text as the body. The
blank line separating the summary from the body is critical (unless
you omit the body entirely); various tools like `log`, `shortlog`
and `rebase` can get confused if you run the two together.

Explain the problem that this commit is solving. Focus on why you
are making this change as opposed to how (the code explains that).
Are there side effects or other unintuitive consequences of this
change? Here's the place to explain them.

Further paragraphs come after blank lines.

 - Bullet points are okay, too

 - Typically a hyphen or asterisk is used for the bullet, preceded
   by a single space, with blank lines in between, but conventions
   vary here

If you use an issue tracker, put references to them at the bottom,
like this:

Resolves: #123
See also: #456, #789