Refactor to allow for two robot players
The existing game loop is based on getting a move from the player in
the first stage, then showing the result. If both players are robots
there will be no player input. This commit introduces redirect, reset
and conditional statements to allow escaping of the player input
stage when there are no human players (NB it currently DOES NOT
implement this but sets up the architecture to enable it).

During the refactor also removed the redundant '/saved_name' from the
app and removed the modulo from Game#outcome (which was redundant as
it only converted a negative index to the corresponding positive
index in the @outcomes array, and ties the result to an array of size
3 so limits extension).