Make Game#player1 into a hash containing the name
Let's make Game#player1 return a hash, and we can store the name and the
player's choice in it. Maybe in future it would be worth creating a
Player object, but it seems unnecessary at the moment since it only
needs to store a name and a choice.

I've also rewritten the test for Game.create so that it doesn't rely on
the implementation of any other methods. Don't know if it's strictly a
good idea to be doubling the behaviour of the class being described...