Implemented the first user story:
As a marketeer
So that I can see my name in lights
I would like to register my name before playing an online game

I've also implemented a function to randomise the computer's move
and also take a move as an input from the webapp. By seeding the
random number generator I was able to successfully test the randomise
feautire of the player class.