Added game class
The game class takes a hierarchy on initialisation and creates a
computer player to play against. To play, pass the game the player's
move, this will return `:player1` if the player wins, `:player2` if
the computer wins or `nil` if it is a draw.
The player's and the computers moves can be retrieved through
reader methods.