Add computer class
Adds computer class to use polymorphism to make tests pass. Previously several
methods were needed in the controller to generate and show a computer move. Now
a computer model is injected with the same pattern as player, allowing for easy
showing and comparing of moves.

This step should make also make it easier to add a multiplayer option in the
future.