Refactor model layer
The model layer was refactored to significantly simplify the logic.

Move classes were created for each type of move so that logic for
pitting moves against eachother was properly encapsulated.

Knowledge of rules messages was encapsulated in its own class,
RulesList.

The MovesList is no longer necessary.

The Game class was simplified based on the above changes.