User Story 3
 # As a marketeer
  # So I can play a game
  # I want to choose a shape to play

created instance variable @shape and used sessions to store it with
params[:shape].

This is possible because in the <form> we give to
name="shape". Params are passed in the form of hash. Provide a way for
information to be passed from a user to your application. Params are
used to allow interactivity on a website, for example by allowing form
data to be sent.